refactor(nats): hold the stream retry in one place
Every stream in Kannon needs the same thing when NATS is briefly slow to
come up: a handful of attempts with a doubling delay, honouring
cancellation while it waits (#365). The dispatcher owned that loop, so a
second stream would have meant a second copy of it, free to ride out a
shorter outage than the first.

`utils.ConfigureStream` now holds it, and each caller keeps only its own
`StreamConfig` — which is the half that differs. The dispatcher's tests
move with the loop, since what they exercise is the retrying and not the
sending stream.

Refs #437
